third_party/blink/web_tests/wpt_internal/app-history/navigate-event/navigate-history-back-after-fragment.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  window.onload = t.step_timeout(() => {
    let start_length = history.length;
    location.hash = "#1";
    assert_equals(history.length, start_length + 1);

    appHistory.onnavigate = t.step_func_done(e => {
      assert_true(e.cancelable);
      assert_true(e.canRespond);
      assert_false(e.userInitiated);
      assert_true(e.hashChange);
      assert_equals(new URL(e.destination.url).hash, "");
      assert_true(e.destination.sameDocument);
      assert_equals(e.formData, null);
    });

    history.back();
  }, 0);
}, "history.back() fires the navigate event and sets hashChange when reversing a fragment navigation");
</script>
